+++ src/usecases/userUsecase.js
@@ -1,5 +1,5 @@
 const User = require("../models/userModel");
-const Wallet = require("../usecases/walletUserUsecase");
+const wallet = require("../usecases/walletUserUsecase");
 const createError = require("http-errors");
 const bcrypt = require("bcrypt");
 const jwt = require("../lib/jwt");
@@ -7,13 +7,13 @@ const saltRounds = 10;
 
 async function login(user_mail, user_password) {
   const userFound = await User.findOne({ user_mail });
-  if (!userFound) throw new createError(401, "Invalid data");
+  if (!userFound) throw new createError(401, "Datos invalidos");
 
   const isValidPassword = await bcrypt.compare(
     user_password,
     userFound.user_password
   );
-  if (!isValidPassword) throw new createError(401, "Invalid data");
+  if (!isValidPassword) throw new createError(401, "Datos invalidos");
   //Expirar token
   return {
     token: jwt.sign({ id: userFound._id }),
@@ -26,7 +26,7 @@ async function createUser(data) {
     user_mail: data.user_mail,
   })
   if (userFound) {
-    throw new createError(432, "user already exist");
+    throw new createError(432, "Usuario no existente");
   }
   const newUser = new User(data);
 
@@ -35,7 +35,7 @@ async function createUser(data) {
     return newUser.save();
   })
 
-  const newWallet = await Wallet.createWalletUser(newUser._id);
+  const newWallet = await wallet.createWalletUser(newUser._id);
   await User.findByIdAndUpdate(newUser._id, { walletUser: newWallet._id });
 
 }
@@ -47,12 +47,16 @@ function getAllUser() {
 function getByIdUser(id) {
   const userFound = User.findById(id);
   if (!userFound) {
-    throw new createError(404, "User not found");
+    throw new createError(404, "Usuario no existente");
   }
   return userFound;
 }
 
 function deleteUserById(id) {
+  const user = User.findById(id)
+  if (!user) {
+    throw new createError(404, 'Usuario no existente')
+  }
   return User.findByIdAndDelete(id);
 }
